:recycle: Refactor EmptyActivity tests
