+++ _integration_tests/test/empty_activities/finish_empty_activity.js
@@ -2,9 +2,9 @@
 
 const should = require('should');
 
-const {TestFixtureProvider, ProcessInstanceHandler} = require('../../dist/commonjs/');
+const {TestFixtureProvider, ProcessInstanceHandler} = require('../../dist/commonjs');
 
-describe('Management API: FinishEmptyActivity', () => {
+describe(`Management API: FinishEmptyActivity`, () => {
 
   let processInstanceHandler;
   let testFixtureProvider;
