+++ _integration_tests/test/empty_activities/get_waiting_for_process_model_in_correlation.js
@@ -3,9 +3,9 @@
 const should = require('should');
 const uuid = require('node-uuid');
 
-const {TestFixtureProvider, ProcessInstanceHandler} = require('../../dist/commonjs/');
+const {TestFixtureProvider, ProcessInstanceHandler} = require('../../dist/commonjs');
 
-describe('Management API: GetEmptyActivitiesForProcessModelInCorrelation', () => {
+describe(`Management API: GetEmptyActivitiesForProcessModelInCorrelation`, () => {
 
   let eventAggregator;
   let processInstanceHandler;
@@ -119,7 +119,7 @@ describe('Management API: GetEmptyActivitiesForProcessModelInCorrelation', () =>
     before(async () => {
       // Create a number of ProcessInstances, so we can actually test pagination
       // We will have a grand total of 10 EmptyActivities after this.
-      for (let i = 0; i < 10; i++) {
+      for(let i = 0; i < 10; i++) {
         await processInstanceHandler.startProcessInstanceAndReturnResult(processModelId, correlationIdPaginationTest);
       }
       await processInstanceHandler.waitForProcessInstanceToReachSuspendedTask(correlationIdPaginationTest, processModelId, 10);
@@ -224,22 +224,16 @@ describe('Management API: GetEmptyActivitiesForProcessModelInCorrelation', () =>
       }
     });
 
-    it('should fail to retrieve the correlation\'s EmptyActivities, when the user is forbidden to retrieve it', async () => {
+    it('should return an empty Array, if the user not allowed to access any suspended EmptyActivities', async () => {
 
       const restrictedIdentity = testFixtureProvider.identities.restrictedUser;
+      const emptyActivityList = await testFixtureProvider
+        .managementApiClient
+        .getEmptyActivitiesForProcessModelInCorrelation(restrictedIdentity, processModelId, correlationId);
 
-      try {
-        const emptyActivityList = await testFixtureProvider
-          .managementApiClient
-          .getEmptyActivitiesForProcessModelInCorrelation(restrictedIdentity, processModelId, correlationId);
-
-        should.fail(emptyActivityList, undefined, 'This request should have failed!');
-      } catch (error) {
-        const expectedErrorMessage = /access denied/i;
-        const expectedErrorCode = 403;
-        should(error.message).be.match(expectedErrorMessage);
-        should(error.code).be.match(expectedErrorCode);
-      }
+      should(emptyActivityList).have.property('emptyActivities');
+      should(emptyActivityList.emptyActivities).be.an.instanceOf(Array);
+      should(emptyActivityList.emptyActivities).have.a.lengthOf(0);
     });
   });
 });
